@@ -195,109 +198,117 @@
     "redux-immutable": "^4.0.0",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.2.0",
-    "remotedev-serialize": "^0.1.0",
-    "request": "^2.83.0",
-    "rm": "^0.1.8",
+    "remotedev-serialize": "^0.1.1",
+    "request": "^2.85.0",
     "secure-random-string": "^1.1.0",
     "sha1": "^1.1.1",
     "sqlite3": "^3.1.13",
-    "ssh2-sftp-client": "^2.0.1",
-    "strong-soap": "^1.5.0",
-    "styled-components": "^3.1.5",
+    "ssh2": "^0.6.0",
+    "ssh2-sftp-client": "^2.1.1",
+    "strong-soap": "^1.7.0",
+    "styled-components": "^3.2.6",
     "then-request": "^5.0.0",
     "transit-immutable-js": "^0.7.0",
-    "transit-js": "^0.8.846",
+    "transit-js": "^0.8.861",
     "v8": "^0.1.0",
     "velocity-animate": "^1.5.0",
     "velocity-react": "^1.3.3",
     "victory": "^0.23.0",
-    "winston": "^2.4.0",
+    "winston": "^2.4.2",
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
     "@types/bcrypt": "^1.0.0",
-    "@types/chai": "^4.1.2",
+    "@types/chai": "^4.1.3",
     "@types/classnames": "^2.2.3",
     "@types/codemirror": "^0.0.43",
     "@types/d3": "^4.7.0",
     "@types/double-ended-queue": "^2.1.0",
+    "@types/downloadjs": "^1.4.0",
     "@types/elasticsearch": "^5.0.22",
-    "@types/jest": "^22.1.3",
+    "@types/graphlib": "^2.1.4",
+    "@types/jest": "^22.2.3",
     "@types/jquery": "^2.0.41",
-    "@types/koa": "2.0.39",
-    "@types/koa-bodyparser": "^3.0.26",
+    "@types/koa": "^2.0.45",
+    "@types/koa-bodyparser": "^4.2.0",
     "@types/koa-favicon": "^2.0.19",
     "@types/koa-logger": "^3.1.0",
-    "@types/koa-passport": "^2.0.4",
+    "@types/koa-passport": "^4.0.1",
     "@types/koa-router": "^7.0.22",
     "@types/koa-send": "^4.1.1",
-    "@types/lodash": "^4.14.104",
+    "@types/lodash": "^4.14.107",
     "@types/loglevel": "^1.5.3",
-    "@types/mysql": "^2.15.2",
-    "@types/node": "8.9.3",
+    "@types/mysql": "^2.15.4",
+    "@types/node": "^9.6.6",
     "@types/node-schedule": "^1.2.1",
-    "@types/passport-local": "^1.0.30",
-    "@types/pg": "7.4.2",
+    "@types/papaparse": "^4.1.33",
+    "@types/passport": "^0.4.5",
+    "@types/passport-local": "^1.0.33",
+    "@types/pg": "^7.4.6",
     "@types/puppeteer": "^0.13.9",
     "@types/rc-slider": "^8.2.3",
     "@types/react": "^16.0.38",
     "@types/react-beautiful-dnd": "^4.0.0",
-    "@types/react-day-picker": "^5.3.0",
     "@types/react-dnd": "^2.0.34",
-    "@types/react-dom": "^16.0.4",
+    "@types/react-dom": "^16.0.5",
     "@types/react-modal": "^1.6.6",
     "@types/react-onclickoutside": "^6.0.3",
     "@types/react-router": "3.0.2",
     "@types/redux-actions": "^1.2.3",
     "@types/request": "^2.47.0",
-    "@types/sinon": "^4.1.3",
-    "@types/sqlite3": "^3.1.1",
+    "@types/sinon": "^4.3.1",
+    "@types/sqlite3": "^3.1.3",
+    "@types/ssh2": "^0.5.35",
     "@types/supertest": "^2.0.4",
-    "@types/tape": "^4.2.31",
-    "@types/winston": "^2.3.8",
+    "@types/tape": "^4.2.32",
+    "@types/winston": "^2.3.9",
     "babel-cli": "^6.26.0",
-    "babel-loader": "^7.1.2",
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.4",
     "babel-plugin-transform-async-to-generator": "^6.22.0",
     "babel-plugin-transform-react-constant-elements": "^6.5.0",
     "babel-plugin-transform-react-display-name": "^6.25.0",
     "babel-plugin-transform-react-inline-elements": "^6.6.5",
     "babel-plugin-transform-regenerator": "^6.26.0",
     "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-0": "^6.24.1",
-    "css-loader": "^0.28.9",
+    "css-loader": "^0.28.11",
     "enzyme": "^2.9.1",
-    "expose-loader": "^0.7.4",
-    "file-loader": "^1.1.6",
-    "fork-ts-checker-webpack-plugin": "^0.3.0",
-    "hard-source-webpack-plugin": "^0.5.18",
+    "expose-loader": "^0.7.5",
+    "file-loader": "^1.1.11",
+    "fork-ts-checker-webpack-plugin": "^0.4.1",
+    "hard-source-webpack-plugin": "^0.6.4",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^22.4.2",
+    "jest": "^22.4.3",
     "json-loader": "^0.5.4",
-    "less-loader": "^0.6.2",
-    "nodemon": "^1.15.1",
+    "less": "^3.0.2",
+    "less-loader": "^4.1.0",
+    "nodemon": "^1.17.3",
     "pre-commit": "^1.2.2",
     "puppeteer": "^0.13.0",
     "raw-loader": "^0.5.1",
     "react-test-renderer": "^15.6.1",
     "redux-mock-store": "^1.5.1",
     "sleep": "^5.1.1",
-    "string-replace-loader": "^1.0.5",
-    "style-loader": "^0.19.1",
+    "string-replace-loader": "^2.1.1",
+    "style-loader": "^0.21.0",
     "supertest": "^3.0.0",
     "svg-react-loader": "^0.4.5",
     "tape": "^4.4.0",
     "thread-loader": "^1.1.2",
-    "ts-jest": "^22.4.0",
-    "ts-loader": "^3.5.0",
-    "ts-node": "^4.1.0",
+    "ts-jest": "^22.4.4",
+    "ts-loader": "^4.2.0",
+    "ts-node": "^6.0.0",
+    "tsconfig-paths": "^3.3.1",
     "tslint": "^5.9.1",
-    "typescript": "^2.7.2",
-    "typescript-formatter": "^7.1.0",
-    "url-loader": "^0.6.2",
-    "webpack": "^3.11.0",
-    "webpack-dev-server": "^2.11.1",
+    "typescript": "^2.8.3",
+    "typescript-formatter": "^7.2.0",
+    "url-loader": "^1.0.1",
+    "webpack": "^4.6.0",
+    "webpack-cli": "^2.0.15",
+    "webpack-dev-server": "^3.1.3",
     "yarn-run-all": "^3.1.1"
   }
 }
