@@ -120,6 +120,7 @@ export interface Props
   hideLine?: boolean;
   canDelete?: boolean;
   showWarning?: boolean;
+  useTooltip?: boolean;
   warningMessage?: string;
   onDelete?: () => void;
   onToggleOpen?: (open: boolean) => void;
@@ -157,7 +158,7 @@ export class RouteSelector extends TerrainComponent<Props>
     return !_.isEqual(this.props, nextProps) || !_.isEqual(this.state, nextState);
   }
 
-  public componentWillMount()
+  public componentDidMount()
   {
     // if (this.props.optionSets && !this.props.optionSets.get(0).hideSampleData)
     // {
@@ -293,6 +294,7 @@ export class RouteSelector extends TerrainComponent<Props>
                     noBorder={true}
                     forceFloat={optionSet.forceFloat}
                     extendRight={true}
+                    useTooltip={this.props.useTooltip}
                   />
               }
               {/*getValueRef={this.handleValueRef}
