@@ -145,7 +145,7 @@ describe('Replay a builder action', () =>
     {
       await takeBuilderActionScreenshot(page);
     });
-  }, 300000);
+  }, 600000);
 
   afterAll(async () =>
   {
