@@ -111,9 +111,11 @@ export const PathfinderText = {
   },
 
   // More section
-  moreSectionTitle: 'Advanced',
+  moreSectionTitle: 'More',
   moreSectionSubtitle:
     'Add advanced functionality to your algorithm, like grouping by a field, adding scripts, and adding personalization algorithms',
+  sizeTitle: 'Number of Results',
+
   // Collapse
   collapseTitle: 'Group by',
   collapseTootlip: 'Force algorithm to only show the top result with matching values for this field',
