@@ -156,21 +156,48 @@ export class FilterUtils
 
     const indexFilters: Block[] = [];
     const otherFilters: Block[] = [];
+    let dummyFilters: Block[] = [];
+
     filterRows.map((row: Block) =>
     {
       if (row.boolQuery === 'filter' && row.filterOp === '=' && row.field === '_index')
       {
         indexFilters.push(row);
+      } else if (row.boolQuery === 'filter' && row.filterOp === 'exists' && row.field === '_id')
+      {
+        dummyFilters.push(row);
       } else
       {
         otherFilters.push(row);
       }
     });
+
+    // one dummy filter is enough
+    if (dummyFilters.length > 1)
+    {
+      dummyFilters = [dummyFilters[0]];
+    }
+    // check current dummyFilters in the card
+    if (dummyFilters.length === 0 && block['dummyFilters'].size > 0)
+    {
+      const dummyFilter = block['dummyFilters'].get(0);
+      if (dummyFilter.field === '_id' && dummyFilter.boolQuery === 'filter' && dummyFilter.key === 'exists')
+      {
+        dummyFilters.push(dummyFilter);
+      } else
+      {
+        dummyFilters.push(
+          dummyFilter.set('field', '_id')
+            .set('boolQuery', 'filter')
+            .set('key', 'exists'),
+        );
+      }
+    }
     // regroup the filter rows first because a new other-filter row added into
     // the index filter list or the type filter list
     block = block.set('indexFilters', Immutable.List(indexFilters));
+    block = block.set('dummyFilters', Immutable.List(dummyFilters));
     block = block.set('otherFilters', Immutable.List(otherFilters));
-
     // update the cached currentIndex;
     let indexField = '';
     if (block['indexFilters'].size > 0)
@@ -223,8 +250,10 @@ export const elasticFilter = _card({
   currentIndex: '',
   // caching the type
   currentType: '',
-  // filters divided as index/other filters
+  // filters divided as index/dummy/other filters
   indexFilters: List([]),
+  // dummyFilters are for ensuring soft bools soft.
+  dummyFilters: List([]),
   otherFilters: List([]),
   cards: Immutable.List([]),
   getChildOptions: FilterUtils.BoolQueryCard.getChildOptions,
@@ -241,7 +270,7 @@ export const elasticFilter = _card({
     colors: getCardColors('filter', Colors().builder.cards.structureClause),
     preview: (c: Card) =>
     {
-      return String(c['indexFilters'].size + c['otherFilters'].size + c['cards'].size) + ' Filters';
+      return String(c['indexFilters'].size + c['dummyFilters'].size + c['otherFilters'].size + c['cards'].size) + ' Filters';
     },
     // this tql is same as tql of other clause cards.
     tql: (block, tqlTranslationFn, tqlConfig) =>
@@ -288,11 +317,28 @@ export const elasticFilter = _card({
                   },
                   {
                     displayType: DisplayType.TEXT,
-                    key: 'value',
+                    key: 'field',
                     getAutoTerms: (schemaState, builderState) =>
                     {
                       return ElasticBlockHelpers.autocompleteMatches(schemaState, builderState, AutocompleteMatchType.Index);
                     },
+                  },
+                ],
+            },
+        },
+        {
+          displayType: DisplayType.ROWS,
+          key: 'dummyFilters',
+          english: 'Soft Bool',
+          factoryType: 'elasticFilterBlock',
+          row:
+            {
+              inner:
+                [
+                  {
+                    displayType: DisplayType.LABEL,
+                    label: 'Soft Filter',
+                    key: null,
                   },
                 ],
             },
